Add default form style for Spree front end
This commit adds a default form with the same style as Spree's front
end, to persist `Spree::PresentNote` records.

The form can be used as it is, or overridden by the application
consuming this gem.

The form has been conceived to be used in the checkout flow, ideally
adding it below the address forms.